DO NOT MERGE: Disabling avahi for coral
When avahi comes up it does hostname discovery
and as part of discovery it looks for unique hostname.
In a network where more then 100 nodes are there,the discovery
process takes a long time.

as avahi uses mDNS which is incompatible with unicast
DNS zones (.local).We should not use avahi in that
network setup.

More info is at the following link
https://web.archive.org/web/20150228130344/http://avahi.org/wiki/AvahiAndUnicastDotLocal

Change-Id: If74f084bcef95f6b9880e218e10a393202849de8
Signed-off-by: Ratan Gupta <ratagupt@in.ibm.com>