WA: Manually add groups and permission
There was a change to add these new groups and permissions to
the code base but since this information is stored in /etc/group,
it is not automatically updated since that file has been written
to and a copy is out in /var/persist/etc/group. If you factory
reset your system then you will get the correct settings but can
not require a factory reset to existing users.

This commit just ensures the new groups are created and root is
in the appropriate groups before starting the user management
service.

Change-Id: I1b35de76daa5d35c85b58983321c2ffd34b33c79
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>