nsswitch: set ldap config as default
Previously the filesystem would have both a default config file, which
works with the 'compat' mode, and the ldap config. We would switch
between these configs based on whether an ldap server is
configured/removed. The nsswitch modes are documented in 'man
nsswitch.conf'.

This was causing an unexpected cache to happen in phosphor-rest-server,
because it would read the default config file before an LDAP server is
configured, and would not read the ldap config again. From 'man
nsswitch.conf' : Within each process that uses nsswitch.conf, the entire
file is read only once.  If the file is later changed, the process will
continue using the old configuration.

What causes the rest server to read the nsswitch conf would be
invocation of any of the functions documented in the manual, for eg
gethostbyname, via the GNU C library. It's for this reason that
restarting the rest server would cause it pick the ldap config correctly.

Hence, set the ldap enabled config as the default. This config still
searches for users, groups and passwords in local files (such as
/etc/group) and then tries to contact an ldap server, if there is one.
So there shouldn't be any major side effects of setting this config file
as the default.

Change-Id: Ifaab4ab75a6fb6303f52fa76654a29505e96b23c
Signed-off-by: Deepak Kodihalli <dkodihal@in.ibm.com>

Conflicts:
	meta-phosphor/recipes-core/base-files/base-files_%.bbappend