synclist: Only sync network configuration and uuid
There's a scenario where LDAP is enabled and the root user is
disabled. In this scenario the disabling of the root user would be
synced as part of syncing the password/shadow files, but not all the
LDAP app configuration files are synced, leaving the BMC locked out
if there is a chip switch.
To avoid issues, just sync the network configuration information so
that if the BMC is setup with a static IP that the user still has
easy access to it, and the machine id (uuid) which is used by systemd
on startup.

Change-Id: Ie3d944a581081e7faccb499e57a9a86d205c5123
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>