recipes-phosphor: Fix systemd unit dependences of ipmid, mboxd
With the implementation of DBus and IPMI transports for the HIOMAP
protocol, mboxd now depends on ipmid being alive in its shutdown
sequence in order to propagate BMC status events to the host firmware.

Mark phosphor-ipmi-host.service as depending on
org.openbmc.HostIpmi.service and mboxd as depending on
phosphor-ipmi-host.service to ensure the mboxd state update
event on SIGTERM is propagated to the host before the services are torn
down.

Change-Id: Iab4bd4d0d88ad8775f5c6593bb3f651cf4230ff6
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>