p10bmc: Add I2CDevice to Rainier VPD JSON
Add the xyz.openbmc_project.Inventory.Decorator.I2CDevice interface to
the extraInterfaces sections for the Rainier PCIe cards.  This contains
the bus and address of the VPD EEPROM.  Note that that a cable card
cannot be plugged into the C6 slot so not entry was added there.

This will be used by entity-manager, which has the ability to pull in
properties from interfaces on the same object path as it probed on.  As
cable cards will be probed using the CC keyword, it will then be able to
substitute in these properties when it sees $bus and $address in its
JSON files.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>