Disable the old Power/Thermal
In downstream we have enabled the new PowerSubsystem and ThermalSubsystem.
Let's turn off the old Power/Thermal, so it is not used.
The old Power / Thermal doesn't have working powercapping so would
have to remove that, just disabling the old Power / Thermal schemas
is more in line with our direction so just do that instead.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>