phosphor-hostfw-image: Fix duplicate content
The image-hostfw was being created in the ${S} directory which is
preserved. Therefore, the mksquashfs command would duplicate files that
already existed in that image, example:
80a00001.lid
80a00001.lid_1
80a00001.lid_2

The import of the hostfw binaries is setup to take place in the ${B}
directory which is cleaned up on every run, therefore the image-hostfw
should be created on that directory as well, the copy it over to the
${S} directory for packaging.

Tested: Just re-building the phosphor-hostfw-image recipe with this
change, it reduced the size of the ${S}/image-hostfw file
(tmp/work/all-openbmc-linux/phosphor-hostfw-image/1.0-r1/image-hostfw)
from 233MB to 97MB. Mounted this new image and verified no duplicate
files *_N existed.

Change-Id: If09afb60b316e1f1b60f6a5b9e1b0057865f0b58
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>