Remove hypervisor network dbus objects from Settings
There is an intermediate app developed in the networkd
for hypervisor network configuration. The dbus object path
will be "/xyz/openbmc_project/network/hypervisor/eth<0/1>".
The dbus objects with the same path have been hosted by
xyz.openbmc_project.Settings service which was used earlier
by pldm and bmcweb.
With the new app, the latter is not needed and the default
configuration should be removed.

This pr removes the default hypervisor n/w configuration.

Fixes the defect at:
https://w3.rchland.ibm.com/projects/bestquest/?defect=SW539017

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I57f936b54ef95dd6ef91d615e426748ed5ad0b2e