meta-ibm: Manually magic close the fan watchdog
To support the watchdog application being able to restart after expiring
the watchdog (by being killed) multiple times per BMC reboot, a manual
magic close must be issued by writing a 'V' into the watchdog device
before starting the watchdog.  Without that, the underlying hardware
watchdog wouldn't actually start running when the watchdog app was
started again, such as on the next boot.

Most watchdogs simply reboot the system upon expiration, but in this
case it's being used to trip a GPIO that the fan hardware latches on to
go to full speed so must support being restarted multiple times.

The busybox watchdog application used to support this, but it was
removed since it apparently caused some undesireable print statements
from the kernel on other platforms.

An alternative may be to fix this in the ASPEED watchdog driver, though
I'm not sure what all that would entail or if it even makes sense.

Tested: Expire and start the watchdog multiple times and the fans now go
to full speed each time the watchdog expires, unlike before.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I770c4fd7cca15949abd7de9dc07975562080f032