Expire the admin password (#1878)
This changes the admin account so its password is expired.  The password
must be changed before access to the admin account is granted.

The password can be changed via the Redfish API (see test below).
Other ways to change the password include via the `login admin` command
password change dialog, or by logging in as root and running the `paswd
admin` command.

This does not change the password itself which remains 0penBmc.

The password change is only needed one-time per BMC.  Existing BMCs
upgrading to this will not need to change their passsword.  There is
nothing on the BMC which would cause any password to expire, that is,
no password aging.

Tested:
On a newly installed BMC the admin account password is
administratively expired.

Changing the password via the `login admin` command works.

Changing the password via the Redfish API works.  Specifically,
you can login via POST /redfish/v1/SessionService/Sessions and get
the PasswordChangeRequired message in the response.  From that state
you can PATCH the new password into
/redfish/v1/AccountService/Accounts/admin which change the password.

I did not test the SSH password change dialog because I believe it
is not present, and is not a primary use case.

The password is preserved across rebooting the BMC.

I did not test the pasword is preserved accross BMC code update.
I did not test if the password is reset by BMC factory reset.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

Co-authored-by: Joseph Reynolds <joseph-reynolds@charter.net>