meta-ibm: p10bmc: Add opkg to RDEPENDS:${PN}-extras (#1906)
Providing `opkg` in the image allows us to easily install packages
after-the-fact, usually to debug systems in the lab.

This is not a mechanism for providing firmware updates.

No specific configuration is supplied to enable its use, and hosting of
package archives is left as an exercise to the user. That said, bitbake
generates its own package archive in the course of creating the output
image. It also provides a `package-index` target, which after execution,
allows you to expose the archive with no more work than:

```
$ python3 -m http.server --directory tmp/deploy/ipk 8080`
```

The addition of opkg with the chosen options increases the uncompressed
filesystem size by 3855360 bytes (~3.68MiB).

Change-Id: I7731a4ebe08249e2216e8ca1984f8e8d87558687
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
(cherry picked from commit 14d942bd35c9d5c2b3cd0ef12a038aaf0a7f3ed7)

Co-authored-by: Andrew Jeffery <andrew@aj.id.au>