meta-ibm:vpd-json: Add Slot iface to pcie cards
Entity manager needs the slot number of PCIe cards on D-Bus so it can
use it when building up names for the sensors on the cards.

For example, with:
   "Name": "NVMe JBOF Card $slotnumber Temp",

There would be a sensor name of:
    /xyz/openbmc_project/temperature/NVMe_JBOF_Card_10_Temp

for the temp sensor on the card in slot C10.  Otherwise, it's just
numbered in the order it's discovered which isn't stable.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: Ia1d1dab5e47cd844d8dd2bcff8a21fa9535d85cd