meta-aspeed:kernel: remove everest phr function (#1964)
A lengthy discussion has started with the Everest system owner and
pinhole reset logic circuit designer. The net is that it's still not
clear if this function behaves as expected on Everest and it needs more
investigation.

Until a solution is found, remove the pinhole reset logic on Everest
systems by removing the dts definition. Userspace does nothing if it can
not find the GPIO by name.

This PR uses a fork because there is one new commit in the linux ghe
repo that has a co-req with a pdbg change that is still in debug. This
branch we're pointing to just brings in the one fix we need for Everest.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>

Co-authored-by: Andrew Geissler <geissonator@yahoo.com>