ibm:p10bmc:assoc: Add cable card sensor assocs
Add inventory associations for the TMP435 sensors on the Flett
(NVMe_JBOF_Card_Cx_Temp) and Bear River/Lake (PCIe_Cable_Card_Cx_Temp) cards.

These associations allow the sensor-monitor code to find the FRU callout
if the sensors were to assert threshold alarms.

It's possible the system configs don't actually allow the Bear* cards to
be in every single slot, but this commit is erring on the side of
caution and adds them to all slots in case some rules change later.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I4e61a2fcfe608ec43f68cae03a48be8192eb94b5