Disable root login (#1886)
This disables the root account.  The root account password is locked and
the root account is expired.  The root acount is removed from user
management, so it no longer shows up on the D-Bus interface or anywhere
that uses that (such as in Redfish APIs or the Web GUI).

The service account can still use the sudo command to become the
effective root user.

Tested:
Root logins fail.
Sudo still works.
The root account no longer appears under the phosphor user manager.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

Co-authored-by: Joseph Reynolds <joseph-reynolds@charter.net>