meta-ibm: inventory: Remove AssetTag Dependency
By default, the PIM build picks up a assettag YAML that causes
PIM to reset the AssetTag property on the "/system" inventory
object to "" every time it starts.

For P10BMC images, the AssetTag is to be persisted across BMC boots
and this action gets in the way of doing that.

Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>