Don't expire the root account
This removes a failed attempt to create the root account as an expired
account.  Note: the account was *not* expired due to a code bug which was
diagnosed during the build process (user 1970-01-01 doesn't exist, unable
to modify it), but did not fail to build.  And now it is not desired to
expire the root account.  Note the root password is locked.

Tested:
Builds without the error message.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>