hostfw: Add squashfs options for reproducible builds
The mksquashfs command adds data to the header such as the creation
time, this makes the generated squashfs image not match the one
generated by the build. Use additional options offered by the squashfs
tools for reproducible builds such as setting the timestamp to 0.

Tested: Verified the generated image had the fields such as timestamp
set to 0s.

Change-Id: I7f6a9e599878d78ce5a3d180d816aa9994442985
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>