phosphor-ipmi-net: disable by default on p10bmc
By default, disable the IPMI network interface on the first boot of a
p10bmc system after a factory reset. Once disabled on that first boot,
ensure this service does not run again so we honor the customer setting
from there on out.

Need to ensure that the IPMI network services are started for a brief
time so that they show up in the systemd service list. The
service-config-manager, which provides the external API to Redfish (via
bmcweb) for enabling and disabling the service requires the service show
up in the systemd list so that it knows to provide the enable/disable
mechanism.

Initially tried having this new service utilize the
service-config-manager D-Bus API's to disable but service-config-manager
requires all services in multi-user.target to have started before
looking for services to control. So, ended up with a chicken-egg
situation with that implementation. Chose instead to utilize systemctl
commands to explicitly stop the IPMI network services and socket.

This commit feels like a workaround to a service-config-manager
deficiency. https://github.com/openbmc/service-config-manager/issues/6
was opened upstream to hopefully track a more user-friendly solution to
this problem.

Tested:
- Verified in SIMICS that on the first boot of a fresh image, the IPMI
network services were not running
- Verified after above test, that Redfish API could be used to enable
the IPMI network services and that they started
- Verified that after a BMC reboot, the IPMI network services were still
running
- Verified that when IPMI services were then disabled via Redfish API,
that they were stopped and were not restarted on a BMC reboot

Signed-off-by: Andrew Geissler <andrewg@us.ibm.com>