Updating the backing build to hostboot to pick up new functionality
This commit updates the hostfw_trace.py script to fix issues with
global variables not being cleared which can cause errors or warnings
to be propagated and will lead to confusion.

This commit also picks up an update to the setup.py which will
have update the version of the peltool.

CQ: SW552597
Signed-off-by: Roland Veloz <rveloz@us.ibm.com>